fix: Long → Integer 타입 명시 오류 수정

+++ src/main/java/site/courseregistrationsystem/enrollment/application/EnrollmentService.java
@@ -39,7 +39,7 @@ public class EnrollmentService {
 		return new EnrolledLecture(savedEnrollment.fetchLectureId());
 	}
 
-	public EnrolledLecture enrollLectureByNumber(Long studentPk, Long lectureNumber) {
+	public EnrolledLecture enrollLectureByNumber(Long studentPk, Integer lectureNumber) {
 		Student student = studentRepository.findById(studentPk).orElseThrow(NonexistenceStudentException::new);
 		Lecture lecture = lectureRepository.findByNumberWithSchedule(lectureNumber)
 			.orElseThrow(NonexistenceLectureException::new);
